Move java-3/ to java/
