--- app/java-3/src/main/java/com/google/cloudclientapi/CensusApplication.java
@@ -1,12 +1,0 @@
-package com.google.cloudclientapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CensusApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(CensusApplication.class, args);
-  }
-}
