--- app/java-3/src/main/java/com/google/cloudclientapi/SquirrelSegment.java
@@ -1,17 +1,0 @@
-package com.google.cloudclientapi;
-
-/*
- * A "segment" is a unique combination of a specific fur, age, and location.
- */
-class SquirrelSegment {
-
-  // Variables names may not follow Java conventions because these names need
-  // to match names used for the other versions/languages (e.g. Python)
-  // of this Jump Start Solution and the names originally used in the raw data file (CSV).
-  public int _count = 0;
-  public int Chasing = 0;
-  public int Climbing = 0;
-  public int Eating = 0;
-  public int Foraging = 0;
-  public int Running = 0;
-}
