--- app/java-3/src/test/java/com/google/cloudclientapi/SquirrelCensusDictionaryBuilderTest.java
@@ -1,11 +1,0 @@
-package com.google.cloudclientapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SquirrelCensusDictionaryBuilderTest {
-
-  @Test
-  void contextLoads() {}
-}
