+++ app/java/project.toml
@@ -14,4 +14,4 @@
 
 [[build.env]]
   name = "GOOGLE_RUNTIME_VERSION"
-  value = "18"
+  value = "17"
