+++ app/java/src/main/java/com/google/cloudclientapi/SquirrelSegment.java
@@ -16,13 +16,18 @@
 
 package com.google.cloudclientapi;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
+/*
+ * A "segment" is a unique combination of a specific fur, age, and location.
+ */
+class SquirrelSegment {
 
-@SpringBootApplication
-public class CensusApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(CensusApplication.class, args);
-  }
+  // Variables names may not follow Java conventions because these names need
+  // to match names used for the other versions/languages (e.g. Python)
+  // of this Jump Start Solution and the names originally used in the raw data file (CSV).
+  public int _count = 0;
+  public int Chasing = 0;
+  public int Climbing = 0;
+  public int Eating = 0;
+  public int Foraging = 0;
+  public int Running = 0;
 }
