+++ app/java/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-#  Copyright 2021 Google LLC
+#  Copyright 2024 Google LLC
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -15,3 +15,5 @@
 server.port=${PORT:8080}
 
 spring.thymeleaf.cache=false
+
+spring.application.name=cloud-client-api
