--- app/java/src/main/resources/logback-spring.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-Copyright 2021 Google LLC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-
-<!-- Learn more about Spring GCP configuration:
-https://cloud.spring.io/spring-cloud-gcp/multi/multi__stackdriver_logging.html -->
-<configuration>
-  <appender name="CONSOLE_JSON" class="ch.qos.logback.core.ConsoleAppender">
-      <encoder class="ch.qos.logback.core.encoder.LayoutWrappingEncoder">
-          <layout class="org.springframework.cloud.gcp.logging.StackdriverJsonLayout">
-              <includeTraceId>true</includeTraceId>
-              <includeSpanId>true</includeSpanId>
-              <includeLevel>true</includeLevel>
-              <includeThreadName>false</includeThreadName>
-              <includeLoggerName>false</includeLoggerName>
-              <includeContextName>false</includeContextName>
-              <includeMDC>true</includeMDC>
-              <!-- Uncomment below for local development only -->
-              <!-- <jsonFormatter class="ch.qos.logback.contrib.jackson.JacksonJsonFormatter">
-                  <prettyPrint>true</prettyPrint>
-              </jsonFormatter> -->
-          </layout>
-      </encoder>
-  </appender>
-  <root level="INFO">
-    <appender-ref ref="CONSOLE_JSON" />
-  </root>
-</configuration>
